Add source-repository setting to cabal file
I was told that Hackage uses this to display a link to the source
repository (not relevant right now since this code is not published
on hackage), and there may be a Cabal version soon which has a command
to automatically clone the repository.

So: not very important, but it doesn't hurt.