Added Pipe module
My first Haskell module; just as an exercise, define a little
type-safety wrapper around TChan which models a chan as a source/sink
pair. The idea is that I want to enforce unidirectional communication
at compile time by passing a chan handle to a thread which can be only
used for reading/writing.